# parser: accelerate DecodingLayerParser