@notti 2019-06-10
Test results on my laptop (Core i5-5300U)
```
name                                    old time/op  new time/op  delta
DecodingLayerParserIgnorePanic-4         138ns ± 0%   136ns ± 0%   -1.81%  (p=0.000 n=8+10)
DecodingLayerParserHandlePanic-4         192ns ± 0%   187ns ± 1%   -2.71%  (p=0.000 n=9+9)
DecodingLayerParserSparseIgnorePanic-4   138ns ± 0%   125ns ± 0%   -9.42%  (p=0.000 n=8+8)
DecodingLayerParserSparseHandlePanic-4   192ns ± 0%   175ns ± 1%   -9.06%  (p=0.000 n=9+9)
DecodingLayerParserArrayIgnorePanic-4    138ns ± 0%   131ns ± 0%   -5.07%  (p=0.000 n=8+7)
DecodingLayerParserArrayHandlePanic-4    192ns ± 0%   181ns ± 2%   -6.00%  (p=0.000 n=9+10)
DecodingLayerParserMapIgnorePanic-4      138ns ± 0%   137ns ± 1%   -0.80%  (p=0.002 n=8+10)
DecodingLayerParserMapHandlePanic-4      192ns ± 0%   190ns ± 1%   -1.43%  (p=0.000 n=9+10)
DecodingLayerArray-4                     138ns ± 0%   128ns ± 0%   -7.25%  (p=0.002 n=8+10)
DecodingLayerMap-4                       138ns ± 0%   133ns ± 2%   -3.62%  (p=0.000 n=8+10)
DecodingLayerSparse-4                    138ns ± 0%   114ns ± 0%  -17.03%  (p=0.000 n=8+10)
```

(all compared to the old implementation of DLP; last 3 compared to old implementation of DecodingLayerParserIgnorePanic)