@notti 2019-06-10
Hmm right now this looks like a lot code duplication: `LayersDecoder` looks the same in every layer. Can this be somehow reduced to one implementation? Otherwise, this looks good to me.