@yerden 2019-06-11
@notti thanks for the feedback. Unfortunately, as I stated earlier, attempt to make a single `LayersDecoder` function and use it by reference in any DecodingLayerContainer implementation makes performance gain much less palpable in new containers. And even, in case of DecodingLayerMap, this leads to some performance loss. The reason, I believe, is inability to inline `Decoder()` method inside indrect `LayersDecoder()` function call during compilation. I'll take a look about what can be done here.