@yerden 2019-06-11
@notti thank you for your feedback. Unfortunately, I cannot fully avoid code duplication because any attempt to indirectly call some `LayersDecoder()` in every container implementation makes performance gain much less palpable. And even, in case of DecodingLayerMap, it causes some performance loss.

Though, I've managed to reduce code duplication a little in 627aaa6.