@notti 2019-06-17
Hmm there is still 5 times the same function. I think there is no other way without loosing inlining/other optimizations. So last resort I could think of: how about generating the code like with gen2.go in layers?