@yerden 2019-06-20
Windows tests were failing, I closed and reopened PR to trigger new build and they are fine now. A mystery.